feat(iam): improve disable credentials checks
