+++ prowler/config/config.yaml
@@ -1,5 +1,11 @@
 # AWS Configuration
 aws:
+  # AWS IAM Configuration
+  # aws.iam_user_accesskey_unused
+  max_unused_access_keys_days: 30
+  # aws.iam_user_console_access_unused
+  max_console_access_days: 30
+
   # AWS EC2 Configuration
   # aws.ec2_elastic_ip_shodan
   shodan_api_key: null
